+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @authhero/demo
 
+## 0.5.11
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies [a0a18c9]
+  - authhero@0.32.0
+  - @authhero/kysely-adapter@0.26.1
+
 ## 0.5.10
 
 ### Patch Changes